css modified; test code added; reorganized code
