+++ tethysapp/data_rods_explorer/model_objects.py
@@ -1,5 +1,7 @@
 from socket import gethostname
-import inspect
+from os import path
+from datetime import datetime, timedelta
+
 
 if 'apps.hydroshare' in gethostname():
     GEOSERVER_URL = 'http://apps.hydroshare.org:8181/geoserver/wms'
@@ -22,74 +24,11 @@ DATARODS_PNG = ('http://giovanni.gsfc.nasa.gov/giovanni/daac-bin/wms_ag4?VERSION
                 '&bbox={0},{1},{2},{3}')  # -119,30,-107,36
 
 
-def parse_model_database_from_file():
-    db_file = inspect.getfile(inspect.currentframe()).replace('model_objects.py',
-                                                              'public/data/model_config.txt')
-    new_model_switch = False
-    model_options = []
-    var_dict = {}
-    wms_vars = {}
-    datarods_tsb = {}
-    with open(db_file, mode='r') as f:
-        f.readline()  # skip column headings line
-        f.readline()
-        for line in f.readlines():
-            if line == '\n':
-                new_model_switch = True
-                continue
-            line = line.strip()
-            linevals = line.split('|')
-            if new_model_switch:
-                model_vals = linevals[0].split('~')
-                model_name = model_vals[0]
-                model_key = model_vals[1]
-                datarods_tsb[model_key] = model_vals[4]
-                model_options.append((model_name, model_key))
-                new_model_switch = False
-                continue
-            else:
-                model_key = linevals[0]
+def init_model():
+    global MODEL_OPTIONS, VAR_DICT, WMS_VARS, DATARODS_TSB, MODEL_FENCES
 
-                if model_key not in wms_vars:
-                    wms_vars[model_key] = {}
-
-                wms_vars[model_key][linevals[1]] = [linevals[2], linevals[3], linevals[4]]
-
-                if model_key not in var_dict:
-                    var_dict[model_key] = []
-
-                var_dict[model_key].append({
-                    "text": "%s %s" % (linevals[3], linevals[4]),
-                    "value": linevals[1],
-                    "layerName": linevals[5]
-                })
-
-    return model_options, var_dict, wms_vars, datarods_tsb
-
-
-def set_model_options(model_options):
-    global MODEL_OPTIONS
-    MODEL_OPTIONS = model_options
-
-
-def set_model_fences(model_fences):
-    global MODEL_FENCES
-    MODEL_FENCES = model_fences
-
-
-def set_var_dict(var_dict):
-    global VAR_DICT
-    VAR_DICT = var_dict
-
-
-def set_wms_vars(wms_vars):
-    global WMS_VARS
-    WMS_VARS = wms_vars
-
-
-def set_datarods_tsb(datarods_tsb):
-    global DATARODS_TSB
-    DATARODS_TSB = datarods_tsb
+    MODEL_OPTIONS, VAR_DICT, WMS_VARS, DATARODS_TSB = parse_model_database_from_file()
+    MODEL_FENCES = parse_fences_from_file()
 
 
 def get_model_options():
@@ -130,3 +69,89 @@ def get_geoserver_url():
 def get_datarods_png():
     global DATARODS_PNG
     return DATARODS_PNG
+
+
+def parse_fences_from_file():
+    """
+    Get date, time and spatial extent fences. Modelname will default to first one listed in input file.
+    Begin dates will be actual model begin date + 1, and end dates will be actual model end date -1, to avoid chance
+    of user selecting hours out of range for dates that are within range.
+    """
+    model_fences = {}
+
+    fencefile = path.join(path.dirname(path.realpath(__file__)), 'public/data/dates_and_spatial_range.txt')
+
+    with open(fencefile, mode='r') as f:
+        f.readline()  # skip column headings line
+        for line in f.readlines():
+            if not (line == '' or 'Model name' in line):      # end condition
+                line = line.strip()
+                linevals = line.split('|')
+                start_date = (datetime.strptime(linevals[1].split(' ')[0], '%m/%d/%Y') + timedelta(days=1)) \
+                    .strftime('%m/%d/%Y')
+                # begin_time = linevals[1].split(' ')[1]
+                end_date = (datetime.strptime(linevals[2].split(' ')[0], '%m/%d/%Y') - timedelta(days=1)) \
+                    .strftime('%m/%d/%Y')
+                # end_time = linevals[2].split(' ')[1]
+                nbound = linevals[3].split(', ')[0]
+                ebound = linevals[3].split(', ')[1]
+                sbound = linevals[3].split(', ')[2]
+                wbound = linevals[3].split(', ')[3]
+                model_fences[linevals[0]] = {
+                    'start_date': start_date,
+                    'end_date': end_date,
+                    'extents': {
+                        'maxY': nbound,
+                        'maxX': ebound,
+                        'minY': sbound,
+                        'minX': wbound
+                    }
+                }
+
+    return model_fences
+
+
+def parse_model_database_from_file():
+    new_model_switch = False
+    model_options = []
+    var_dict = {}
+    wms_vars = {}
+    datarods_tsb = {}
+
+    db_file = path.join(path.dirname(path.realpath(__file__)), 'public/data/model_config.txt')
+
+    with open(db_file, mode='r') as f:
+        f.readline()  # skip column headings line
+        f.readline()
+        for line in f.readlines():
+            if line == '\n':
+                new_model_switch = True
+                continue
+            line = line.strip()
+            linevals = line.split('|')
+            if new_model_switch:
+                model_vals = linevals[0].split('~')
+                model_name = model_vals[0]
+                model_key = model_vals[1]
+                datarods_tsb[model_key] = model_vals[4]
+                model_options.append((model_name, model_key))
+                new_model_switch = False
+                continue
+            else:
+                model_key = linevals[0]
+
+                if model_key not in wms_vars:
+                    wms_vars[model_key] = {}
+
+                wms_vars[model_key][linevals[1]] = [linevals[2], linevals[3], linevals[4]]
+
+                if model_key not in var_dict:
+                    var_dict[model_key] = []
+
+                var_dict[model_key].append({
+                    "text": "%s %s" % (linevals[3], linevals[4]),
+                    "value": linevals[1],
+                    "layerName": linevals[5]
+                })
+
+    return model_options, var_dict, wms_vars, datarods_tsb
